requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.7.0
astroid==3.3.8
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
fsspec==2024.12.0
gradio==5.9.1
gradio_client==1.5.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
isort==5.13.2
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pylint==3.3.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.8.4
safehttpx==0.1.6
safetensors==0.4.5
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.1
